+++ src/main/java/cz/muni/ics/perunproxyapi/application/facade/ProxyuserFacade.java
@@ -4,11 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.EntityNotFoundException;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.PerunConnectionException;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.PerunUnknownException;
-import cz.muni.ics.perunproxyapi.persistence.models.listOfServices.ServicesDataHolder;
 import cz.muni.ics.perunproxyapi.presentation.DTOModels.UserDTO;
 import lombok.NonNull;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Map;