+++ src/main/java/cz/muni/ics/perunproxyapi/application/service/GuiService.java
@@ -1,9 +1,11 @@
 package cz.muni.ics.perunproxyapi.application.service;
 
-import cz.muni.ics.perunproxyapi.application.facade.parameters.ServicesParams;
+import com.fasterxml.jackson.databind.JsonNode;
+import cz.muni.ics.perunproxyapi.application.facade.parameters.ListOfServicesGuiParams;
+import cz.muni.ics.perunproxyapi.application.facade.parameters.ListOfServicesJsonParams;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.PerunConnectionException;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.PerunUnknownException;
-import cz.muni.ics.perunproxyapi.persistence.models.listOfServices.ServicesDataHolder;
+import cz.muni.ics.perunproxyapi.persistence.models.listOfServices.ListOfServicesDAO;
 import lombok.NonNull;
 
 /**
@@ -16,12 +18,24 @@ public interface GuiService {
     /**
      * Gets all data needed for the list of services.
      *
-     * @param servicesParams Parameters for List of Services.
+     * @param params Parameters for List of Services.
      * @return Gets all data needed for the list of services.
      * @throws PerunUnknownException Thrown as wrapper of unknown exception thrown by Perun interface.
      * @throws PerunConnectionException Thrown when problem with connection to Perun interface occurs.
      */
-    ServicesDataHolder getListOfSps(@NonNull ServicesParams servicesParams)
+    ListOfServicesDAO getListOfSps(@NonNull ListOfServicesGuiParams params)
+            throws PerunUnknownException, PerunConnectionException;
+
+    /**
+     * Gets all data needed for the list of services in JSON.
+     *
+     * @param params Parameters for List of Services.
+     * @return Gets all data needed for the list of services.
+     * @throws PerunUnknownException Thrown as wrapper of unknown exception thrown by Perun interface.
+     * @throws PerunConnectionException Thrown when problem with connection to Perun interface occurs.
+     */
+
+    JsonNode getListOfSpsJson(@NonNull ListOfServicesJsonParams params)
             throws PerunUnknownException, PerunConnectionException;
 
 }