+++ src/main/java/cz/muni/ics/perunproxyapi/presentation/gui/controllers/ListOfSpsController.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import cz.muni.ics.perunproxyapi.application.facade.GuiFacade;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.PerunConnectionException;
 import cz.muni.ics.perunproxyapi.persistence.exceptions.PerunUnknownException;
-import cz.muni.ics.perunproxyapi.persistence.models.listOfServices.ServicesDataHolder;
+import cz.muni.ics.perunproxyapi.persistence.models.listOfServices.ListOfServicesDAO;
 import cz.muni.ics.perunproxyapi.presentation.gui.GuiProperties;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,10 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
-
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 
 import static cz.muni.ics.perunproxyapi.presentation.rest.config.PathConstants.GUI;
 import static cz.muni.ics.perunproxyapi.presentation.rest.config.PathConstants.NO_AUTH_PATH;
@@ -30,7 +27,7 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
  * @author Dominik Baranek <baranek@ics.muni.cz>
  */
 @Controller
-@RequestMapping(value = NO_AUTH_PATH + GUI)
+@RequestMapping(value = NO_AUTH_PATH)
 @Slf4j
 public class ListOfSpsController {
 
@@ -57,9 +54,9 @@ public class ListOfSpsController {
      * @throws PerunConnectionException Thrown when problem with connection to Perun interface occurs.
      * @throws IOException Thrown when error occurs during reading a file.
      */
-    @GetMapping(value = "/services")
+    @GetMapping(value = GUI + "/services")
     public ModelAndView getListOfSps() throws PerunUnknownException, PerunConnectionException, IOException {
-        ServicesDataHolder listOfSps = facade.getListOfSps();
+        ListOfServicesDAO listOfSps = facade.getListOfSps();
 
         ObjectMapper mapper = new ObjectMapper();
 
@@ -74,7 +71,7 @@ public class ListOfSpsController {
         mav.addObject("showStaging", listOfSps.isShowStaging());
         mav.addObject("showProduction", listOfSps.isShowProduction());
 
-        return this.addCommonOptions(mav);
+        return this.addCommonGuiOptions(mav);
     }
 
     /**
@@ -86,23 +83,15 @@ public class ListOfSpsController {
      * @throws PerunConnectionException Thrown when problem with connection to Perun interface occurs.
      */
     @ResponseBody
-    @GetMapping(value = "/services/json", produces = APPLICATION_JSON_VALUE)
-    public Map<String, JsonNode> getListOfSpsJson() throws IOException, PerunUnknownException, PerunConnectionException {
-        ServicesDataHolder services = facade.getListOfSps();
-        ObjectMapper mapper = new ObjectMapper();
-
-        Map<String, JsonNode> propertiesToShow = new HashMap<>();
-        propertiesToShow.put("statistics", mapper.valueToTree(services.getStatistics()));
-        //propertiesToShow.put("services", mapper.valueToTree(services.getServicesJson()));
-
-        return propertiesToShow;
+    @GetMapping(value = "/services", produces = APPLICATION_JSON_VALUE)
+    public JsonNode getListOfSpsJson() throws IOException, PerunUnknownException, PerunConnectionException {
+        return facade.getListOfSpsInJson();
     }
 
-    private ModelAndView addCommonOptions(ModelAndView mav) {
+    private ModelAndView addCommonGuiOptions(ModelAndView mav) {
         mav.addObject(HEADER, guiProperties.getHeaderPath());
         mav.addObject(FOOTER, guiProperties.getFooterPath());
         mav.addObject(LANGUAGE_BAR_ENABLED, guiProperties.isLanguageBarEnabled());
-
         return mav;
     }